Tests: Extracted a common function for fetching performance data.
